refactor: extracting inlineDecoration types into separate file
