+++ src/vs/editor/common/viewModel.ts
@@ -18,6 +18,7 @@ import { BracketGuideOptions, IActiveIndentGuideInfo, IndentGuide } from './text
 import { IViewLineTokens } from './tokens/lineTokens.js';
 import { ViewEventHandler } from './viewEventHandler.js';
 import { VerticalRevealType } from './viewEvents.js';
+import { InlineDecoration, SingleLineInlineDecoration } from './viewModel/inlineDecorations.js';
 
 export interface IViewModel extends ICursorSimpleModel {
 
@@ -395,40 +396,6 @@ export class ViewLineRenderingData {
 			return strings.containsRTL(lineContent);
 		}
 		return false;
-	}
-}
-
-export const enum InlineDecorationType {
-	Regular = 0,
-	Before = 1,
-	After = 2,
-	RegularAffectingLetterSpacing = 3
-}
-
-export class InlineDecoration {
-	constructor(
-		public readonly range: Range,
-		public readonly inlineClassName: string,
-		public readonly type: InlineDecorationType
-	) {
-	}
-}
-
-export class SingleLineInlineDecoration {
-	constructor(
-		public readonly startOffset: number,
-		public readonly endOffset: number,
-		public readonly inlineClassName: string,
-		public readonly inlineClassNameAffectsLetterSpacing: boolean
-	) {
-	}
-
-	toInlineDecoration(lineNumber: number): InlineDecoration {
-		return new InlineDecoration(
-			new Range(lineNumber, this.startOffset + 1, lineNumber, this.endOffset + 1),
-			this.inlineClassName,
-			this.inlineClassNameAffectsLetterSpacing ? InlineDecorationType.RegularAffectingLetterSpacing : InlineDecorationType.Regular
-		);
 	}
 }
