+++ src/vs/editor/common/viewModel/modelLineProjection.ts
@@ -9,7 +9,8 @@ import { IRange } from '../core/range.js';
 import { EndOfLinePreference, ITextModel, PositionAffinity } from '../model.js';
 import { LineInjectedText } from '../textModelEvents.js';
 import { InjectedText, ModelLineProjectionData } from '../modelLineProjectionData.js';
-import { SingleLineInlineDecoration, ViewLineData } from '../viewModel.js';
+import { ViewLineData } from '../viewModel.js';
+import { SingleLineInlineDecoration } from './inlineDecorations.js';
 
 export interface IModelLineProjection {
 	isVisible(): boolean;
